requirements.txt
pandas
xgboost
scikit-learn
mlflow
google-cloud-aiplatform
google-cloud-storage
kfp
db-dtypes